requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv

langchain
langchain-google-genai